Holovibes 10.3

+++ Holovibes/sources/gui/other/gui_curve_plot.cc
@@ -106,7 +106,7 @@ void CurvePlot::load_data_vector()
     if (!data_vect_.empty())
     {
         size_t copied_elts_nb = data_vect_.fill_array(chart_vector_, points_nb_);
-        new_data.reserve(copied_elts_nb);
+        new_data.reserve(static_cast<int>(copied_elts_nb));
 
         ++auto_scale_curr_points_;
 
@@ -131,10 +131,9 @@ void CurvePlot::auto_scale()
 {
     std::vector<ChartPoint> tmp = chart_vector_;
 
-    auto minmax = std::minmax_element(tmp.cbegin(),
-                                      tmp.cend(),
-                                      [&](const ChartPoint& lhs, const ChartPoint& rhs)
-                                      { return curve_get_(lhs) < curve_get_(rhs); });
+    auto minmax = std::minmax_element(tmp.cbegin(), tmp.cend(), [&](const ChartPoint& lhs, const ChartPoint& rhs) {
+        return curve_get_(lhs) < curve_get_(rhs);
+    });
 
     double min = curve_get_(*(minmax.first));
     double max = curve_get_(*(minmax.second));
